Owendev

+++ MicdUp/server/database/publicSchema/mutations/user.js
@@ -142,33 +142,24 @@ const deleteAccount = {
     };
     try {
       // For each person that you are following, unfollow them
-      for (let i = 0; i < context.profile.following.length; i++) {
-        const followingProfile = Profile.findById(context.profile.following[i]);
-        const index = followingProfile.followers.findIndex((id) => {
-          return id.toString() === context.profile.id;
-        });
-        // const index = followingProfile.followers[context.profile.id];
-        if (index > -1) {
-          followingProfile.followers.splice(index, 1);
-          await followingProfile.save({ session });
-        }
+      let keys = Array.from(context.profile.following.keys());
+      for (let i = 0; i < keys.length; i++) {
+        const followingProfile = Profile.findById(keys[i]);
+        followingProfile.followers.delete(context.profile.id);
+        await followingProfile.save({ session });
       }
 
       //For each of your followers, have them unfollow you
-      for (let i = 0; i < context.profile.followers.length; i++) {
-        const followerProfile = Profile.findById(context.profile.followers[i]);
-        const index = followingProfile.followers.findIndex((id) => {
-          return id.toString() === context.profile.id;
-        });
-        if (index > -1) {
-          followerProfile.following.splice(index, 1);
-          await followerProfile.save({ session });
-        }
+      keys = Array.from(context.profile.followers.keys());
+      for (let i = 0; i < keys.length; i++) {
+        const followerProfile = Profile.findById(keys[i]);
+        followerProfile.following.delete(context.profile.id);
+        await followerProfile.save({ session });
       }
       // TODO: Force user to confirm delete account.
-      await Post.findByIdAndDelete(context.profile.id);
-      await User.findByIdAndDelete(context.user.id);
-      await Profile.findByIdAndDelete(context.profile.id);
+      await Post.findByIdAndDelete({ owner: context.profile.id }, { session });
+      await User.findByIdAndDelete(context.user.id, { session });
+      await Profile.findByIdAndDelete(context.profile.id, { session });
       await session.commitTransaction();
     } catch (err) {
       returnObject.success = false;
